requirements.txt
openai
python-docx
termcolor
python-dotenv